requirements.txt
discord-components==2.1.2
discord.py==1.7.3